~ | Move private.py's @private_api_auth to api_decorators.py
